Refactor UI and main logic, separate DOM rendering to ui.js
- Moved all DOM rendering logic from main.js to ui.js
- Added functions in ui.js for rendering bookmarks and showing empty messages
- Updated main.js to handle events, state management, and call ui.js rendering functions
- Updated index.html to include empty message placeholders for accessibility
- Ensured accessibility features: aria-labels on buttons, aria-live for like counter, and keyboard navigable
- Preserved previous functionality: like buttons, copy URL buttons, bookmark form submission, and user dropdown integration
- Improved code structure to match project tickets and rubric requirements
