+++ index.html
@@ -14,6 +14,7 @@
         <select id="user-dropdown">
           <option value="">--Select a user--</option>
         </select>
+        <p id="user-empty-message" style="display:none;color:red;"></p>
       </section>
 
       <section id="bookmark-form">
@@ -34,6 +35,7 @@
       <section id="bookmark-list">
         <h2>Bookmarks</h2>
         <ul id="bookmarks-container"></ul>
+        <p id="bookmark-empty-message" style="display:none;color:red;"></p>
       </section>
     </main>
   </body>
